Refactored ParcelCalendar implementation and added colours to collection locations.
